Update README.md
The way I wrote my contributions is basically the PR titles instead of the previous "Updating x". Not sure if we want to stay consistent but there might be a bit of overlap on this one so the specifics could be nice. Just to note, Folders/Files included and Detailed Set Up are just copied from Iteration 4, so they will need to be updated as well. I'm going to at least add to the Detailed Set Up once the sync PR goes through.